requirements.txt
openai
praw
pandas
ipykernel
python-dotenv
pydantic
tabulate